+++ tests/entities/test_field.py
@@ -17,4 +17,5 @@ def test_simple_construct():
     assert field.included == []
     assert field.original_included == []
     assert field.meta == {}
+    assert field.object_meta == {}
     assert field.__repr__() == '<Field: id=123, type=dictionary>'